RFFC5071: Separate hardware-specific code.
Conflicts:
	firmware/CMakeLists.txt
	firmware/hackrf_usb/Makefile
	firmware/mixertx/Makefile
	firmware/sgpio/CMakeLists.txt
	firmware/simpletx/Makefile
	firmware/spiflash/CMakeLists.txt
	firmware/startup/Makefile
	firmware/startup_systick/Makefile
	firmware/startup_systick_perfo/Makefile
